+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - removed the method `is_equivalence` from `Constructor`. You should instead
   use the `is_equivalence` method on the `Rule`.
 
+### Changed
+- the `CartesianProduct` now considers compositions of all parameters and
+  not just `n`.
+- the `RelianceProfile` type changed to work multiple parameters. It is now a
+  dictionary pointing from parameters to the values.
+
 ### Fixed
 - ignore rules where the left and non-empty right hand sides are the same